database works tested with postman

+++ database.py
@@ -1,10 +1,11 @@
-from sqlalchemy import create_engine, Column, Integer, String, DateTime
+from sqlalchemy import create_engine, Column, Integer, String, DateTime, Float
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import sessionmaker
 from sqlalchemy import Integer 
 
 
 DATABASE_URL = "sqlite:///./sensor.db"
+
 engine = create_engine(DATABASE_URL, connect_args={"check_same_thread": False})
 
 
@@ -17,14 +18,14 @@ Base = declarative_base()
 class TemperatureSensor(Base):
     __tablename__ = "temperature_sensor"
     id = Column(Integer, primary_key=True, index=True)
-    value = Column(Integer, nullable=False)
+    value = Column(Float, nullable=False)
     unit = Column(String, nullable=False)
     timestamp = Column(DateTime, nullable=False)
 
 class AirQualitySensorPM25(Base):
     __tablename__ = "air_quality_sensor_pm2_5"
     id = Column(Integer, primary_key=True, index=True)
-    value = Column(Integer, nullable=False)
+    value = Column(Float, nullable=False)
     unit = Column(String, nullable=False)
     timestamp = Column(DateTime, nullable=False)
 
@@ -32,7 +33,7 @@ class AirQualitySensorPM25(Base):
 class LightSensor(Base):
     __tablename__ = "light_sensor"
     id = Column(Integer, primary_key=True, index=True)
-    value = Column(Integer, nullable=False)
+    value = Column(Float, nullable=False)
     unit = Column(String, nullable=False)
     timestamp = Column(DateTime, nullable=False)
